api/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
coloredlogs==15.0.1
dataclasses-json==0.6.3
distro==1.9.0
Django==5.0.1
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
flatbuffers==23.5.26
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
humanfriendly==10.0
idna==3.6
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-openai==0.0.5
langsmith==0.0.85
Markdown==3.5.2
marshmallow==3.20.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
onnxruntime==1.17.0
openai==1.10.0
opencv-python==4.9.0.80
packaging==23.2
Pillow==10.0.1
pinecone-client==3.0.2
protobuf==4.25.2
pyclipper==1.3.0.post5
pydantic==2.6.0
pydantic_core==2.16.1
pypdf==4.0.1
python-dotenv==1.0.1
pytz==2023.4
PyYAML==6.0.1
rapidocr-onnxruntime==1.3.10
regex==2023.12.25
requests==2.31.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==2.2.0
yarl==1.9.4
